pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = "Thesis project"
authors = ["Aleksey Lesnikov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
psycopg2-binary = "^2.9.5"
envparse = "^0.2.0"
gunicorn = "^20.1.0"
djangorestframework-simplejwt = "^5.2.2"
social-auth-app-django = "^5.0.0"
django-filter = "^22.1"


[tool.poetry.group.dev.dependencies]
ansible-vault-win = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
